LayoutTests/canvas/philip/tests/2d.imageData.create2.round-expected.txt

< [index] >
2d.imageData.create2.round
createImageData(w, h) is rounded the same as getImageData(0, 0, w, h)
References: 2d.imageData.createround Bugzilla
Actual output:
Passed
